split the vec discretization in dedicated files
