+++ src/pygeon/__init__.py
@@ -18,14 +18,12 @@ from pygeon.discretizations.discretization import Discretization
 from pygeon.discretizations.vec_discretization import VecDiscretization
 
 from pygeon.discretizations.fem.hcurl import Nedelec0, Nedelec1
-from pygeon.discretizations.fem.hdiv import RT0, BDM1, VecBDM1, VecRT0
-from pygeon.discretizations.fem.h1 import Lagrange1, VecLagrange1
-from pygeon.discretizations.fem.l2 import (
-    PwConstants,
-    PwLinears,
-    VecPwConstants,
-    VecPwLinears,
-)
+from pygeon.discretizations.fem.hdiv import RT0, BDM1
+from pygeon.discretizations.fem.h1 import Lagrange1
+from pygeon.discretizations.fem.l2 import PwConstants, PwLinears
+from pygeon.discretizations.fem.vec_hdiv import VecBDM1, VecRT0
+from pygeon.discretizations.fem.vec_h1 import VecLagrange1
+from pygeon.discretizations.fem.vec_l2 import VecPwConstants, VecPwLinears
 
 from pygeon.discretizations.vem.hdiv import VRT0, VBDM1, VecVRT0
 from pygeon.discretizations.vem.h1 import VLagrange1, VecVLagrange1
